# src/api_service/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API с JWT авторизацией",
    "description": "Пример API, использующего JWT для авторизации в Swagger UI",
    "version": "1.0"
  },
  "paths": {
...
    "/v1/bookings/get": {
      "get": {
        "tags": [
          "Booking"
        ],
        "summary": "Get Booking",
        "operationId": "get_booking_v1_bookings_get_get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "booking_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booking Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDTO"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticate",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Not authenticate"
                }
              }
            }
          },
          "404": {
            "description": "Booking not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Booking not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/bookings/list/user": {
      "get": {
        "tags": [
          "Booking"
        ],
        "summary": "List User Bookings",
        "operationId": "list_user_bookings_v1_bookings_list_user_get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1,
                  "examples": [
                    1522105862
                  ]
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingDTO"
                  },
                  "title": "Response List User Bookings V1 Bookings List User Get"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticate",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Not authenticate"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Access denied"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/bookings/list/coworking": {
      "get": {
        "tags": [
          "Booking"
        ],
        "summary": "List Coworking Bookings",
        "operationId": "list_coworking_bookings_v1_bookings_list_coworking_get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "coworking_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Coworking Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingDTO"
                  },
                  "title": "Response List Coworking Bookings V1 Bookings List Coworking Get"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticate",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Not authenticate"
                }
              }
            }
          },
          "404": {
            "description": "Coworking not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Coworking not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/bookings/list/all": {
      "get": {
        "tags": [
          "Booking"
        ],
        "summary": "Get All Bookings",
        "operationId": "get_all_bookings_v1_bookings_list_all_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BookingDTO"
                  },
                  "type": "array",
                  "title": "Response Get All Bookings V1 Bookings List All Get"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticate",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Not authenticate"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Access denied"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/bookings/add": {
      "post": {
        "tags": [
          "Booking"
        ],
        "summary": "Add Bookings",
        "operationId": "add_bookings_v1_bookings_add_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingCreateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Access denied"
                }
              }
            }
          },
          "404": {
            "description": "Booking not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Booking not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/bookings/update": {
      "patch": {
        "tags": [
          "Booking"
        ],
        "summary": "Update Bookings",
        "operationId": "update_bookings_v1_bookings_update_patch",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "booking_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booking Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingCreateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Booking is not waiting",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Booking is not waiting"
                }
              }
            }
          },
          "404": {
            "description": "Booking not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Booking not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/bookings/delete": {
      "delete": {
        "tags": [
          "Booking"
        ],
        "summary": "Delete Bookings",
        "operationId": "delete_bookings_v1_bookings_delete_delete",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "booking_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Booking Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Access denied"
                }
              }
            }
          },
          "404": {
            "description": "Booking not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Booking not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/bookings/activate": {
      "post": {
        "tags": [
          "Booking"
        ],
        "summary": "Activate Booking",
        "operationId": "activate_booking_v1_bookings_activate_post",
        "parameters": [
          {
            "name": "booking_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booking Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Access denied"
                }
              }
            }
          },
          "404": {
            "description": "Booking not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Booking not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/bookings/pending": {
      "get": {
        "tags": [
          "Booking"
        ],
        "summary": "Pending Bookings",
        "operationId": "pending_bookings_v1_bookings_pending_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BookingDTO"
                  },
                  "type": "array",
                  "title": "Response Pending Bookings V1 Bookings Pending Get"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Access denied"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BookingCreateDTO": {
        "properties": {
          "workplaces": {
            "items": {
              "type": "string",
              "format": "uuid4",
              "examples": [
                "57f25aa0-ea44-4aa6-892f-1c3ebf1cab1b"
              ]
            },
            "type": "array",
            "title": "Workplaces"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time",
            "examples": [
              "2025-03-04T06:37:03.959954"
            ]
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "title": "End Time",
            "examples": [
              "2025-03-04T06:37:03.959993"
            ]
          }
        },
        "type": "object",
        "required": [
          "workplaces",
          "start_time",
          "end_time"
        ],
        "title": "BookingCreateDTO"
      },
      "BookingDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid4",
            "title": "Id",
            "examples": [
              "f3047cd8-56e6-46e4-ac2d-757550c1f62a"
            ]
          },
          "user_id": {
            "type": "integer",
            "minimum": 1,
            "title": "User Id",
            "examples": [
              1522105862
            ]
          },
          "workplaces": {
            "items": {
              "$ref": "#/components/schemas/WorkplaceDTO"
            },
            "type": "array",
            "title": "Workplaces"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time",
            "examples": [
              "2025-03-04T06:37:03.959954"
            ]
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "title": "End Time",
            "examples": [
              "2025-03-04T06:37:03.959993"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          },
          "total_price": {
            "type": "integer",
            "minimum": 0,
            "title": "Total Price",
            "examples": [
              1500
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "examples": [
              "2025-03-04T06:37:03.960093"
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "workplaces",
          "start_time",
          "end_time",
          "status",
          "total_price",
          "created_at"
        ],
        "title": "BookingDTO"
      },
      "BookingStatus": {
        "type": "string",
        "enum": [
          "WAITING",
          "PROCESSING",
          "FINISHED"
        ],
        "title": "BookingStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT токен для авторизации. Формат: Bearer <токен>"
      }
    }
  },
  "servers": [
    {
      "url": "/api"
    }
  ]
}